fix: rename things from EmptyActivity to Mutsu
not everything was changed because it wouldn't let me change everything, but there was a lot that changed. I ran the app after making these changes and it ran properly.
