+++ app/src/androidTest/java/com/example/mutsu/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.emptyactivity
+package com.example.mutsu
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
