+++ app/src/main/java/com/example/mutsu/RecipeListScreen.kt
@@ -1,9 +1,5 @@
-package com.example.emptyactivity
+package com.example.mutsu
 
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -15,7 +11,6 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.paddingFromBaseline
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
@@ -26,7 +21,6 @@ import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -34,14 +28,10 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.AccountCircle
 import androidx.compose.material.icons.filled.Favorite
 import androidx.compose.material.icons.filled.Home
 import androidx.compose.material3.CardDefaults
@@ -58,7 +48,7 @@ import androidx.compose.material3.windowsizeclass.WindowSizeClass
 import androidx.compose.material3.windowsizeclass.WindowWidthSizeClass
 import androidx.compose.material3.Scaffold
 import androidx.compose.foundation.layout.Box
-import com.example.emptyactivity.ui.theme.EmptyActivityTheme
+import com.example.mutsu.ui.theme.MutsuTheme
 
 data class Recipe(val name: String)
 
@@ -181,7 +171,7 @@ fun RecipeList(recipeList: List<Recipe>) {
 //Version of the app that uses the horizontal navigation bar
 @Composable
 fun CookBookPortrait(){
-    EmptyActivityTheme {
+    MutsuTheme {
         Scaffold(
             bottomBar = { BottomNavigation() }
         ) { padding ->
@@ -224,7 +214,7 @@ fun HomeScreen(modifier: Modifier = Modifier){
 //Version of the app that uses the vertical navigation bar
 @Composable
 fun CookBookLandscape(){
-    EmptyActivityTheme {
+    MutsuTheme {
         Surface(color = MaterialTheme.colorScheme.background){
             Row{
                 NavigationRail()
