+++ app/src/main/java/com/example/mutsu/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.emptyactivity.ui.theme
+package com.example.mutsu.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
