+++ app/src/test/java/com/example/mutsu/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.emptyactivity
+package com.example.mutsu
 
 import org.junit.Test
